+++ responsive_data_grid/lib/src/filters/int.dart
@@ -41,7 +41,8 @@ class DataGridIntColumnFilterState<TItem extends Object>
   late IntFilterRules filterRules;
 
   void applyCriteria() {
-    if (logic == null || (iValue == null && iValue2 == null)) {
+    if (logic == null ||
+        (gridLogicRequiresValue(logic!) && iValue == null && iValue2 == null)) {
       filterRules.criteria = null;
     } else {
       List<int> values = List<int>.empty(growable: true);
@@ -91,40 +92,7 @@ class DataGridIntColumnFilterState<TItem extends Object>
       children: [
         DropdownButton<Logic?>(
           isExpanded: true,
-          items: [
-            DropdownMenuItem<Logic?>(
-              value: null,
-              child: Text(GridLocalizations.of(context).any),
-            ),
-            DropdownMenuItem<Logic?>(
-              value: Logic.greaterThan,
-              child: Text(Logic.greaterThan.toString()),
-            ),
-            DropdownMenuItem<Logic?>(
-              value: Logic.greaterThanOrEqualTo,
-              child: Text(Logic.greaterThanOrEqualTo.toString()),
-            ),
-            DropdownMenuItem<Logic?>(
-              value: Logic.lessThan,
-              child: Text(Logic.lessThan.toString()),
-            ),
-            DropdownMenuItem<Logic?>(
-              value: Logic.lessThanOrEqualTo,
-              child: Text(Logic.lessThanOrEqualTo.toString()),
-            ),
-            DropdownMenuItem<Logic?>(
-              value: Logic.between,
-              child: Text(Logic.between.toString()),
-            ),
-            DropdownMenuItem<Logic?>(
-              value: Logic.equals,
-              child: Text(Logic.equals.toString()),
-            ),
-            DropdownMenuItem<Logic?>(
-              value: Logic.notEqual,
-              child: Text(Logic.notEqual.toString()),
-            ),
-          ],
+          items: gridFilterLogicItems(context, gridComparableFilterLogics),
           value: logic,
           onChanged: (Logic? value) {
             setState(() {